refactor(circulation): simpan dan hitung uang sebagai rupiah bulat
Rupiah tidak punya satuan pecahan yang dipakai — sen sudah lama tidak beredar
dan tarif denda perpustakaan selalu bilangan bulat. Menyimpannya sebagai
DECIMAL(12,2) berarti setiap nilai melewati konversi float di PHP, dan `sum()`
atas kolom desimal mengembalikan string yang mudah tercampur saat dijumlahkan
lagi di laporan — tempat kesalahan pembulatan menumpuk tanpa terlihat.

- kolom `fines.amount` dan `return_transactions.fine_amount` menjadi
  unsignedBigInteger, dengan cast integer pada modelnya
- FineAmountCalculator dan OperationalRules::fineDailyAmount() bertipe int
- halaman Aturan Operasional menolak tarif pecahan, sehingga nilai bukan bulat
  tidak pernah masuk lewat pintu depan

Migrasinya memeriksa lebih dulu apakah ada nilai pecahan yang sudah tersimpan,
dan BERHENTI bila ada alih-alih membulatkannya diam-diam. Membulatkan uang
tanpa sepengetahuan siapa pun adalah cara kehilangan kepercayaan pada angka.

MoneyIsWholeRupiahTest memeriksa tipe kolom di basis data, bukan hanya di PHP,
serta membuktikan denda 365 hari dan penjumlahan lintas denda tetap tepat.

727 test hijau di SQLite dan MySQL; Pint dan Larastan bersih.

+++ app/Modules/Circulation/Support/FineAmountCalculator.php
@@ -5,10 +5,14 @@ namespace App\Modules\Circulation\Support;
 /**
  * Perkalian hari keterlambatan dengan tarif harian — dan hanya itu.
  * Tarifnya datang sebagai argumen dari OperationalRules.
+ *
+ * Seluruhnya bilangan bulat: rupiah tidak punya satuan pecahan yang dipakai,
+ * dan menjaganya tetap integer berarti tidak ada nilai uang yang pernah
+ * melewati float — tempat kesalahan pembulatan menumpuk tanpa terlihat.
  */
 class FineAmountCalculator
 {
-    public static function calculate(int $lateDays, float $perDayRate): float
+    public static function calculate(int $lateDays, int $perDayRate): int
     {
         return max(0, $lateDays) * max(0, $perDayRate);
     }
